Do not modify the original `authorized_keys`
overwrite it instead with the generated one.

Google adds managed keys and annotates them with `# Added by Google`
comment. Previously, when the script sorts the file, the annotation goes
at the top, thus on reboot, the first entry is removed from the file by
Google

With this change we generate the content of the file on a temporary
location and then move it to `~/.ssh/authorized_keys`

Co-authored-by: Danail Branekov <danailster@gmail.com>