Allow non-sudo global npm installations
eirini-station's provision-user.sh now configures npm to install
globally into $HOME/.npm-packages. This changes sets up PATH and MANPATH
to use that new location.

This is all because npm is a big dodgy and prior to the version shipped
with nodejs-v19 can corrupt files in /usr/bin. So it really should never
be invoked as root.